% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bd_albums}
\alias{bd_albums}
\title{Names, year of publication, song, and song position for each Bob Dylan Album}
\format{A data.frame with 1064 rows and 4 columns
\describe{
  \item{Albums}{Album name}
  \item{Year}{Year of publication}
  \item{Song}{Name of Song}
  \item{Position}{Ordinal position of song on album}
  }
 @source \url{http://www.bobdylan.com/}}
\usage{
bd_albums
}
\description{
Names, year of publication, song, and song position for each Bob Dylan Album
}
\keyword{datasets}
