% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bd_concerts}
\alias{bd_concerts}
\title{Information about time, place and set list for Bob Dylan Concerts}
\format{A data.frame with 54542 rows and 7 columns
\describe{
  \item{Date}{Date of concert}
  \item{Venue}{Name of venue}
  \item{City}{City location of venue}
  \item{State/Country}{The state (in in the USA) or country where venue is located}
  \item{Songs}{The name of a song played at the concert}
  \item{Position}{The ordinal position of the song in the set list}
  \item{OriginalLocation}{The unseparated City/State(Country) from the raw data}
  }}
\source{
\url{http://www.bobdylan.com/}
}
\usage{
bd_concerts
}
\description{
Information about time, place and set list for Bob Dylan Concerts
}
\keyword{datasets}
