% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bd_songs}
\alias{bd_songs}
\title{Names and Lyrics of Bob Dylan Songs}
\format{A data.frame with 634 rows and 3 variables
\describe{
  \item{SongId}{Song id number}
  \item{Song}{Song name}
  \item{Lyrics}{A vector with line by line lyrics for the song}
  \item{Stems}{A vector of unique stems found in the song}
  \item{StemsByLine}{A list of vectors of stems, one per line of the song}
  }}
\source{
\url{http://www.bobdylan.com/}
}
\usage{
bd_songs
}
\description{
A dataset containing the names and lyrics of over 600 Bob Dylan songs. Very incomplete.
}
\keyword{datasets}
