% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyric_functions.R
\name{find_bd_lyrics}
\alias{find_bd_lyrics}
\title{Search for Bob Dylan Songs containing a given lyrical snippet}
\usage{
find_bd_lyrics(lyrics)
}
\arguments{
\item{lyrics}{a string of lyrics to search for}
}
\value{
a subset of bd_songs containing songs with at least one match with lyrics
}
\description{
Search for Bob Dylan Songs containing a given lyrical snippet
}
\seealso{
Other lyric_functions: \code{\link{find_line_matches}},
  \code{\link{find_stems}}, \code{\link{get_word_stems}}
}
