% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyric_functions.R
\name{find_line_matches}
\alias{find_line_matches}
\title{Given a lyric snippet and a song name, return vector of matches by line}
\usage{
find_line_matches(song, lyrics, return_values = FALSE)
}
\arguments{
\item{song}{String name of a song in bd_songs$Song}

\item{lyrics}{String of lyrics to search}

\item{return_values}{if TRUE return list of stem matches, otherwise return a vector of counts of matches}
}
\value{
a vector of match counts or a list of matching stems
}
\description{
Given a lyric snippet and a song name, return vector of matches by line
}
\seealso{
Other lyric_functions: \code{\link{find_bd_lyrics}},
  \code{\link{find_stems}}, \code{\link{get_stem_match}},
  \code{\link{get_word_stems}},
  \code{\link{match_all_stems}}
}
