% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyric_functions.R
\name{find_stems}
\alias{find_stems}
\title{Find which stems in a given source text can be found in a given target text}
\usage{
find_stems(stems_to_search, stems_to_search_in)
}
\description{
Find which stems in a given source text can be found in a given target text
}
\seealso{
Other lyric_functions: \code{\link{find_bd_lyrics}},
  \code{\link{find_line_matches}},
  \code{\link{get_word_stems}}
}
