% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyric_functions.R
\name{get_word_stems}
\alias{get_word_stems}
\title{Convert a sentence, or list of sentences, into a bag of unique stems}
\usage{
get_word_stems(sentence)
}
\arguments{
\item{sentence}{a space separated string of words, or a list of the same}
}
\value{
a vector of unique stems found in sentence
}
\description{
For a provided text, this function will:
 - convert the sentence to a vector of words
 - convert each word to lower case
 - replace ’ with '
 - remove all punctuation other that '
 - replace -in' with -ing
 - remove snowball stop words (using tidy_text stopwords)
 - stem the words with SnowballC::wordStem
 - convert to a unique vector
}
\seealso{
Other lyric_functions: \code{\link{find_bd_lyrics}},
  \code{\link{find_line_matches}}, \code{\link{find_stems}}
}
