% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyric_functions.R
\name{match_all_stems}
\alias{match_all_stems}
\title{Find all Bob Dylan songs containing given stems}
\usage{
match_all_stems(stems)
}
\arguments{
\item{stems}{a vector of stems, as generated by \code{get_word_stems()}}
}
\value{
a data.frame with names 'f' and 'n'
}
\description{
Returns a data.frame() of song names and hit counts
}
\seealso{
Other lyric_functions: \code{\link{find_bd_lyrics}},
  \code{\link{find_line_matches}},
  \code{\link{find_stems}}, \code{\link{get_stem_match}},
  \code{\link{get_word_stems}}
}
